move send_to_inbox to Inbox.
